+++ Main/Shared/Models/IEntity.cs
@@ -6,11 +6,11 @@ namespace AsciiPinyin.Web.Shared.Models;
 [JsonDerivedType(typeof(Chachar))]
 public interface IEntity
 {
-    /*
-     * The string type must be used even for single characters.
-     * The char type tends to malfunction when sent over HTTP requests.
-     */
+    // The string type must be used even for single characters.
+    // The char type tends to malfunction when sent over HTTP requests.
     string? TheCharacter { get; set; }
 
-    byte? Strokes { get; set; }
+    // Must use signed number types, because blazor client-side model validation doesn't work with unsigned numeric types.
+    // The 'sbyte' would be enough in all situations, but the blazor client-side model validation doesn't work with it => must use 'short'.
+    short? Strokes { get; set; }
 }
